requirements.txt
aiogram==3.0.0
asyncpg==0.28.0
sqlalchemy==2.0.19
alembic==1.11.3
openpyxl==3.1.2
redis==5.0.0
pandas==2.0.3
plotly==5.16.1
python-dotenv==1.0.0
pydantic-settings==2.0.3
pytest-asyncio==0.21.1
pytest-dotenv==0.5.2
pytest-async-sqlalchemy==0.2.0
apscheduler==3.10.4
apscheduler-di==0.1.0
ruff==0.0.287